WIP support for modular Spring AOT
This commit which leverages
https://github.com/sdeleuze/spring-framework/tree/modular-spring-aot
introduces support for modular Spring AOT, with the main goal to
provide more control on which benefits/side effects are enabled.

It introduces 4 new boolean properties to the Gradle AOT tasks:
 - beanRegistration
 - predefinedClasses
 - classpathIndexes
 - reachabilityMetadata
When none of those 4 boolean properties are defined, they are all
enabled to keep backward compatibility.

Build configuration example:

 processAot {
 	predefinedClasses = true
 	classpathIndexes = true
 }

 Other possible options for DevXP:
  - Some artifact types enabled by default on the JVM when Spring AOT processing happen
    but -Dspring.aot.enabled is not specified
  - Have some kind of "JVM mode" flag